Refactor pages components to have individual folder structure
