+++ server/client/src/pages/Projects/index.jsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from 'react';
 
-import { Table } from '../components/Table';
-import { fakeAuthAPI } from '../auth/fakeAuthAPI';
+import { Table } from './Table';
+import { fakeAuthAPI } from '../../auth/fakeAuthAPI';
 
 export default function Projects() {
   const [projects, setProjects] = useState([]);
@@ -9,13 +9,12 @@ export default function Projects() {
   useEffect(() => {
     async function fetchData() {
       const projects = await fakeAuthAPI.get_projects();
-      console.log(projects);
       setProjects(projects);
     }
     fetchData();
   }, []);
   
   return (
-    <Table projectsList={projects}></Table>
+    <Table data={projects}></Table>
   )
 }
