+++ server/client/src/pages/index.jsx
@@ -1,0 +1,16 @@
+import Home from './Home';
+import SignUp from './SignUp';
+import SignIn from './SignIn';
+import Org from './Org';
+import Projects from './Projects';
+import Issues from './Issues';
+
+
+export {
+  Home,
+  SignUp,
+  SignIn,
+  Org,
+  Projects,
+  Issues
+};
